Extract DictionaryPopup
